Setting up a project structure using Priler/tgbotbase from GitHub, adding 'send_sticker' method to the '/start' command
