+++ config.py
@@ -1,4 +1,9 @@
 
+"""
+Contains the config
+"""
+
 import os
 
+# get environment variable
 TOKEN = os.environ.get('BOT_TOKEN')
