+++ handlers/callbacks.py
@@ -1,0 +1,8 @@
+
+"""
+Handle events when clicking on buttons
+"""
+
+from aiogram import types
+from dispatcher import dp
+import config
